modified API and directory structure for the local implementation of the triangular solver algorithm
