+++ test/unit/CMakeLists.txt
@@ -21,4 +21,4 @@ add_subdirectory(communication)
 add_subdirectory(factorization)
 add_subdirectory(matrix)
 add_subdirectory(memory)
-add_subdirectory(mc)
+add_subdirectory(solver)
