--- test/unit/mc/CMakeLists.txt
@@ -1,11 +1,0 @@
-#
-#Distributed Linear Algebra with Future(DLAF)
-#
-#Copyright(c) 2018 - 2019, ETH Zurich
-#All rights reserved.
-#
-#Please, refer to the LICENSE file in the root directory.
-#SPDX - License - Identifier : BSD - 3 - Clause
-#
-
-DLAF_addTest(test_triangular_solve_local SOURCES test_triangular_solve_local.cpp USE_MAIN HPX)
